Relocate  folders
